apps/webhook_server/pyproject.toml
[project]
name = "sharpedge-webhooks"
version = "0.1.0"
description = "Stripe webhook server for SharpEdge"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "stripe>=11.0",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "sharpedge-shared",
    "sharpedge-db",
    "sharpedge-odds",
    "pillow>=12.1.1",
    "firebase-admin>=6.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
sharpedge-webhooks = "sharpedge_webhooks.main:run"

[tool.uv.sources]
sharpedge-shared = { workspace = true }
sharpedge-db = { workspace = true }
sharpedge-odds = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sharpedge_webhooks"]
